explicit type

+++ HetGNN/code_gcn/graph_augmentation.py
@@ -85,7 +85,7 @@ def het_edge_perturbation_from_prior(
                 new_edge_index.append(sampled_edge_index)
                 new_edge_type.append(sampled_edge_type)
 
-    new_edge_index = torch.cat(new_edge_index, dim=1)
+    new_edge_index = torch.cat(new_edge_index, dim=1).long()
     new_edge_type = torch.cat(new_edge_type).int().view(-1,)
 
     generated_adj_matrix = to_dense_adj(
